+++ crates/bevy_input_focus/Cargo.toml
@@ -71,12 +71,10 @@ libm = ["bevy_math/libm", "bevy_window/libm"]
 [dependencies]
 # bevy
 bevy_app = { path = "../bevy_app", version = "0.18.0-rc.2", default-features = false }
-bevy_camera = { path = "../bevy_camera", version = "0.18.0-rc.2", default-features = false }
 bevy_ecs = { path = "../bevy_ecs", version = "0.18.0-rc.2", default-features = false }
 bevy_input = { path = "../bevy_input", version = "0.18.0-rc.2", default-features = false }
 bevy_math = { path = "../bevy_math", version = "0.18.0-rc.2", default-features = false }
 bevy_picking = { path = "../bevy_picking", version = "0.18.0-rc.2", default-features = false, optional = true }
-bevy_ui = { path = "../bevy_ui", version = "0.18.0-rc.2", default-features = false }
 bevy_window = { path = "../bevy_window", version = "0.18.0-rc.2", default-features = false }
 bevy_reflect = { path = "../bevy_reflect", version = "0.18.0-rc.2", features = [
   "glam",
