+++ crates/bevy_input_focus/src/lib.rs
@@ -23,6 +23,7 @@ extern crate std;
 extern crate alloc;
 
 pub mod directional_navigation;
+pub mod navigator;
 pub mod tab_navigation;
 
 // This module is too small / specific to be exported by the crate,
