+++ crates/bevy_ui/Cargo.toml
@@ -19,6 +19,7 @@ bevy_derive = { path = "../bevy_derive", version = "0.18.0-rc.2" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.18.0-rc.2" }
 bevy_image = { path = "../bevy_image", version = "0.18.0-rc.2" }
 bevy_input = { path = "../bevy_input", version = "0.18.0-rc.2" }
+bevy_input_focus = { path = "../bevy_input_focus", version = "0.18.0-rc.2" }
 bevy_math = { path = "../bevy_math", version = "0.18.0-rc.2" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.18.0-rc.2" }
 bevy_sprite = { path = "../bevy_sprite", version = "0.18.0-rc.2", features = [
